Fix another error in humidex calculation
humidex calculation contains a 2nd math.exp that should not be there.
The revised formula agrees with the humidex calculation in the weewx
wxformulas.py as well as Wikipedia and
http://www.physlink.com/reference/weather.cfm. weewx currently
calculates humidex internally but unless the user saves this to archive
it (currently) is not available in templates. Once weewx issue #13
(https://github.com/weewx/weewx/issues/13) is resolved the humidex
calculation can be removed from this template.